ønsker å laste data bare en gang har flyttet sesjon til topnivå har flyttet index til å videresende til enten enkeltview hvis man har en gjeldende rapportering eller oversikten hvis man ikke har det da jeg ikke ønsker å laste samtlige rapporteringer på hver eneste sidelast
