+++ app/routes/_index.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "@remix-run/router";
 
 export async function loader() {
-  return redirect("/rapp");
+  return redirect("/rapportering");
 }
